Implemented a simple board drawing system.
I'd hoped to use the Unicode chess symbols, but Quil's default font didn't know how to deal with them when I increased the font sizes. It's possible that if I did some font loading, etc., I could get that to work, but for the moment we'll just go with the text representation I have here.