Added atomic state.
I added two atoms (`best-board` and `num-iterations`) and basic hill-climbing code to `location_genome`, and then removed all the search from the GUI. The search and the GUI communicate via these atoms (which the GUI uses to find out the state of the search). I also added the number of iterations to the GUI output so we can see how long the search has been going.

I added a short sleep before the search starts so that the GUI has a chance to "build" and get displayed before the search has raced too far ahead.

This finishes Issue #6.