Added support for Poisson distribution for number of swaps.
Initially the number of swaps in a mutation was taken uniformly in the range [1, 3]; this changes that to 1+x, where x is drawn from the Poisson distribution with lambda=2. The hope was that this would occasionally allow for some bigger jumps, which might help deal with being stuck in dead ends. No idea if it _actually_ helps, though.